fix: correct date-range inputs so they submit properly
The datetime-polyfill wasn't adjusting the date onInput or setting a
default date to use. I fixed both of these bugs so it works well now. I
chose to add event listeners for each part of the datetime input to
avoid co-dependent observers that would end up in an infinte loop.